Stick word list to screen when scrolling puzzle
The issue I am attempting to fix here is that when you have a very large
puzzle, scrolling down to look at the puzzle does not scroll the list of
words, and so the user would have to scroll back up to edit the word
list. This change modifies the word list (only when the screen is wide
enough for it to live on the left side of the puzzle) to stick to the
screen as the user scrolls the puzzle. Hopefully the behavior is
somewhat intuitive.